tests/requirements.txt
pytest
Pillow
numpy
